requirements.txt
telethon==1.31.1
python-dotenv==1.0.0
loguru==0.7.2
aiohttp==3.9.1
pyyaml==6.0.1
psutil==5.9.8
